+++ src/common/i18n/fr_FR.js
@@ -116,6 +116,8 @@ pwiAccounts.i18n = {
                 signin_error: 'L\'adresse de messagerie et/ou le mot de passe sont invalides. Merci de bien vouloir les vérifier.\nEn cas de doute, vous pouvez vous rapprocher de notre administrateur.',
                 signup_error: 'Impossible de créer ce nouveau compte.',
                 signup_success: 'Nouveau compte %s créé',
+                signup_autoconnect: 'Le nouveau compte a été automatiquement connecté',
+                signup_noconnect: 'Le compte n\'est pas connecté',
                 verify_error: 'Humm... Désolé, mais il semble que le jeton utilisé soit expiré.<br />Vous devez en redemander un nouveau.',
                 verify_text: 'Bonjour,<br />Votre adresse de messagerie est maintenant considérée comme vérifiée.<br />'
                     +'Vous pouvez participer aux forums (et en profiter).',