+++ src/server/js/methods.js
@@ -3,6 +3,7 @@
  */
 import { Accounts } from 'meteor/accounts-base';
 
+// make sure we remove all sensitive information before returning to the client
 _cleanUser = function ( user ){
     if( user ){
         if( user.services ){
@@ -22,31 +23,31 @@ Meteor.methods({
 
     // find a user by his email address
     'pwiAccounts.byEmailAddress'( email ){
-        console.log( 'pwiAccounts.byEmailAddress' );
+        //console.debug( 'pwiAccounts.byEmailAddress' );
         return _cleanUser( Accounts.findUserByEmail( email ));
     },
 
     // find a user by his internal (mongo) identifier
     'pwiAccounts.byId'( id ){
-        console.log( 'pwiAccounts.byId' );
+        //console.debug( 'pwiAccounts.byId' );
         return _cleanUser( Meteor.users.findOne({ _id: id }));
     },
 
     // find the user who holds the given reset password token
     'pwiAccounts.byResetToken'( token ){
-        console.log( 'pwiAccounts.byResetToken' );
+        //console.debug( 'pwiAccounts.byResetToken' );
         return _cleanUser( Meteor.users.findOne({ 'services.password.reset.token': token },{ 'services.password.reset': 1 }));
     },
 
     // find a user by his username
     'pwiAccounts.byUsername'( username ){
-        console.log( 'pwiAccounts.byUsername' );
+        //console.debug( 'pwiAccounts.byUsername' );
         return _cleanUser( Accounts.findUserByUsername( username ));
     },
 
     // find the user who holds the given email verification token
     'pwiAccounts.byEmailVerificationToken'( token ){
-        console.log( 'pwiAccounts.byEmailVerificationToken' );
+        //console.debug( 'pwiAccounts.byEmailVerificationToken' );
         return _cleanUser( Meteor.users.findOne({ 'services.email.verificationTokens': { $elemMatch: { token: token }}},{ 'services.email':1 }));
     },
 
@@ -66,5 +67,13 @@ Meteor.methods({
     //  - options
     'pwiAccounts.sendVerificationEmail'( id ){
         return Accounts.sendVerificationEmail( id );
+    },
+
+    'pwiAccounts.sendVerificationEmailByEmail'( email ){
+        const u = Accounts.findUserByEmail( email );
+        //console.debug( u );
+        if( u ){
+            Accounts.sendVerificationEmail( u._id );
+        }
     }
 });